TINY-9946: Convert Enter tests from AccordionPluginTest to webdriver tests
